add: web socket et traitement front

+++ src/main/java/fr/hey/PatataChat/controllers/HomeController.java
@@ -1,12 +1,17 @@
 package fr.hey.PatataChat.controllers;
 
 import fr.hey.PatataChat.dto.UserDto;
+import fr.hey.PatataChat.dto.UserMessageDto;
 import fr.hey.PatataChat.entities.Message;
 import fr.hey.PatataChat.entities.UserEntity;
 import fr.hey.PatataChat.services.MessageService;
 import fr.hey.PatataChat.services.UserService;
 import fr.hey.PatataChat.services.security.IAuthenticationFacade;
+import fr.hey.PatataChat.utils.RoleChecker;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.messaging.handler.annotation.MessageMapping;
+import org.springframework.messaging.handler.annotation.Payload;
+import org.springframework.messaging.handler.annotation.SendTo;
 import org.springframework.stereotype.Controller;
 import org.springframework.util.ObjectUtils;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -37,13 +42,15 @@ public class HomeController {
     public ModelAndView home() {
         // Récupération de l'utilisateur connecté
         UserEntity user = this.authenticationFacade.getUserAuth();
+
         if (!ObjectUtils.isEmpty(user)) {
             ModelAndView modelAndView = new ModelAndView("index");
 
             List<UserDto> users = userService.findAllUsers();
             modelAndView.addObject("users", users);
-            modelAndView.addObject("message",new Message());
-            modelAndView.addObject("messages",messageService.findAllMessagesByDateDesc());
+            modelAndView.addObject("message", new Message());
+            modelAndView.addObject("messages", messageService.findAllMessagesByDateDesc());
+            modelAndView.addObject("isAdmin",RoleChecker.isAdmin(user));
 
             return modelAndView;
         }
@@ -53,23 +60,36 @@ public class HomeController {
 
     }
 
+    @MessageMapping("/chat.sendMessage")
+    @SendTo("/topic/public")
+    public UserMessageDto sendMessage(
+            @Payload UserMessageDto userMessageDTO
+    ) {
+        mapAndSaveMessage(userMessageDTO);
 
-    @PostMapping("/add")
-    public String addMessage(Message message){
-        UserEntity user = authenticationFacade.getUserAuth();
+        return userMessageDTO;
+    }
 
+    private void mapAndSaveMessage(UserMessageDto userMessageDTO) {
+        UserEntity user = userService.findByLogin(userMessageDTO.getUserName());
+        System.out.println("HomeController.mapAndSaveMessage");
+        System.out.println("userMessageDTO = " + userMessageDTO);
+        Message message = new Message();
         message.setUserEntity(user);
         message.setMessageDateTime(OffsetDateTime.now());
+        message.setText(userMessageDTO.getContent());
+
         messageService.createMessage(message);
 
-        return "redirect:/";
     }
 
     @PostMapping("/clear")
-    public String deleteAllMessage(){
+    public String deleteAllMessage() {
         UserEntity user = authenticationFacade.getUserAuth();
 
-       messageService.deleteAllMessages();
+        if (RoleChecker.isAdmin(user)) {
+            messageService.deleteAllMessages();
+        }
 
         return "redirect:/";
     }
